rpi/requirements.txt
loguru==0.7.2
paho-mqtt==2.0.0
python-dotenv==1.0.1
gpiozero==2.0.1
simplejpeg==1.7.2
opencv-python